+++ internal/app/ddnsr53.go
@@ -2,6 +2,9 @@ package app
 
 import (
 	"fmt"
+	"net"
+	"runtime"
+	"strings"
 	"sync/atomic"
 	"time"
 
@@ -10,7 +13,7 @@ import (
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/route53"
 	"github.com/crazy-max/ddns-route53/internal/config"
-	"github.com/crazy-max/ddns-route53/internal/utl"
+	"github.com/crazy-max/ddns-route53/pkg/identme"
 	"github.com/hako/durafmt"
 	"github.com/robfig/cron/v3"
 	"github.com/rs/zerolog/log"
@@ -22,9 +25,10 @@ type Client struct {
 	loc      *time.Location
 	cron     *cron.Cron
 	r53      *route53.Route53
+	im       *identme.Client
 	jobID    cron.EntryID
-	lastIPv4 string
-	lastIPv6 string
+	lastIPv4 net.IP
+	lastIPv6 net.IP
 	locker   uint32
 }
 
@@ -50,6 +54,10 @@ func New(cfg *config.Configuration, loc *time.Location) (*Client, error) {
 			cron.SecondOptional|cron.Minute|cron.Hour|cron.Dom|cron.Month|cron.Dow|cron.Descriptor),
 		)),
 		r53: route53.New(sess, &aws.Config{Credentials: creds}),
+		im: identme.NewClient(
+			fmt.Sprintf("ddns-route53/%s go/%s %s", cfg.App.Version, runtime.Version()[2:], strings.Title(runtime.GOOS)),
+			cfg.Flags.MaxRetries,
+		),
 	}, nil
 }
 
@@ -95,48 +103,49 @@ func (c *Client) Run() {
 			c.cron.Entry(c.jobID).Next)
 	}
 
-	// Get current WAN IP
-	wanIPv4 := ""
+	var wanIPv4 net.IP
 	if c.cfg.Route53.HandleIPv4 {
-		wanIPv4, err = utl.WanIPv4()
+		wanIPv4, err = c.im.IPv4()
 		if err != nil {
-			log.Warn().Err(err).Msg("Cannot get WAN IPv4 address")
+			log.Error().Err(err).Msg("Cannot retrieve WAN IPv4 address")
 		} else {
 			log.Info().Msgf("Current WAN IPv4: %s", wanIPv4)
 		}
-	} else {
-		log.Debug().Msg("")
 	}
 
-	wanIPv6 := ""
+	var wanIPv6 net.IP
 	if c.cfg.Route53.HandleIPv6 {
-		wanIPv6, err = utl.WanIPv6()
+		wanIPv6, err = c.im.IPv6()
 		if err != nil {
-			log.Warn().Err(err).Msg("Cannot get WAN IPv6 address")
+			log.Error().Err(err).Msg("Cannot retrieve WAN IPv6 address")
 		} else {
 			log.Info().Msgf("Current WAN IPv6: %s", wanIPv6)
 		}
 	}
 
-	// Skip if last IP is identical
-	if wanIPv4 == c.lastIPv4 && wanIPv6 == c.lastIPv6 {
-		log.Info().Msg("WAN IPv4/IPv6 addresses have not changed since the last update. Skipping...")
+	if wanIPv4 == nil && wanIPv6 == nil {
+		return
+	}
+
+	// Skip if last IP is identical or empty
+	if wanIPv4.Equal(c.lastIPv4) && wanIPv6.Equal(c.lastIPv6) {
+		log.Info().Msg("WAN IPv4/IPv6 addresses have not changed since last update. Skipping...")
 		return
 	}
 
 	// Create Route53 changes
 	r53Changes := make([]*route53.Change, len(c.cfg.Route53.RecordsSet))
 	for i, rs := range c.cfg.Route53.RecordsSet {
-		if wanIPv4 == "" && rs.Type == route53.RRTypeA {
+		if wanIPv4 == nil && rs.Type == route53.RRTypeA {
 			log.Error().Msgf("No WAN IPv4 address available to update %s record", rs.Name)
 			continue
-		} else if wanIPv6 == "" && rs.Type == route53.RRTypeAaaa {
+		} else if wanIPv6 == nil && rs.Type == route53.RRTypeAaaa {
 			log.Error().Msgf("No WAN IPv6 address available to update %s record", rs.Name)
 			continue
 		}
-		wanIP := wanIPv4
+		recordValue := aws.String(wanIPv4.String())
 		if rs.Type == route53.RRTypeAaaa {
-			wanIP = wanIPv6
+			recordValue = aws.String(wanIPv6.String())
 		}
 		r53Changes[i] = &route53.Change{
 			Action: aws.String("UPSERT"),
@@ -144,7 +153,7 @@ func (c *Client) Run() {
 				Name:            aws.String(rs.Name),
 				Type:            aws.String(rs.Type),
 				TTL:             aws.Int64(rs.TTL),
-				ResourceRecords: []*route53.ResourceRecord{{Value: aws.String(wanIP)}},
+				ResourceRecords: []*route53.ResourceRecord{{Value: recordValue}},
 			},
 		}
 	}